chore: more work on page scroll
