+++ src/utils/quizQuestions.ts
@@ -1,7 +1,5 @@
 import { FluidObject } from "gatsby-image"
-import { IBanners } from "../components/PreSorting/types"
-import { IAnswer, IQuestion } from "../components/SortingQuiz/types"
-import { House } from "./sharedTypes"
+import { House, IAnswer, IBanners, IQuestion } from "../types"
 
 interface IEdges {
   node: {
