--- src/utils/sharedTypes.ts
@@ -1,1 +1,0 @@
-export type House = "CONCH" | "IVY" | "STRATUS"
